+++ openfl-workspace/keras/tensorflow/mnist/src/mnist_utils.py
